+++ ChangeLog.md
@@ -7,10 +7,11 @@
     Release date: 
 
     - Define AccountsCore.Config.get() getter for Accounts configuration options, thus bumping minor candidate version number
-    - Exported AccountsCore.Account class is renamed to AccountsCore.Account
+    - Exported AccountsCore.acAccount class is renamed to AccountsCore.Account
     - Extend Transforms.cleaupUserDocument() to be able to keep some fields
     - Define emailMaxCount(), emailMinCount() and usernameMaxCount(), usernameMinCount() which honor both haveXxx and min/max counts parameters
     - Define applyPublishTransforms(), applyReadTransforms(), applyUpdateTransforms(), updateByQuery() server functions
+    - Make sure transformation functions are only available on server side
 
 ### 2.0.0